refactor(package-config): clone original code
