+++ src/cdk/index.ts
@@ -1,7 +1,8 @@
-export * from './ast';
-export * from './build-component';
-export * from './get-project';
-export * from './html-head-element';
-export * from './project-main-file';
-export * from './project-targets';
-export * from './version-agnostic-typescript';
+export * from './ng-add/package-config';
+export * from './utils/ast';
+export * from './utils/build-component';
+export * from './utils/get-project';
+export * from './utils/html-head-element';
+export * from './utils/project-main-file';
+export * from './utils/project-targets';
+export * from './utils/version-agnostic-typescript';
