+++ src/cdk/utils/html-head-element.ts
