+++ src/cdk/utils/version-agnostic-typescript.ts
