+++ src/material/package-config.ts
@@ -1,5 +1,5 @@
 import { Tree } from '@angular-devkit/schematics';
-import { addPackageToPackageJson as originalAddPackageToPackageJson } from '@angular/cdk/schematics/ng-add/package-config';
+import { addPackageToPackageJson as originalAddPackageToPackageJson } from '../cdk';
 
 /** Adds a package to the package.json in the given host tree. */
 export function addPackageToPackageJson(host: Tree, pkg: string, version: string): Tree {
